+++ core/components/superboxfontawesome/lexicon/en/default.inc.php
@@ -5,7 +5,12 @@
  * @package superboxfontawesome
  * @subpackage lexicon
  */
-
+$_lang['superboxfontawesome.excludeClasses'] = 'Exclude Classes';
+$_lang['superboxfontawesome.excludeClasses_desc'] = 'Comma separated list of excluded FontAwesome icon css classes.';
 $_lang['superboxfontawesome.fontawesome'] = 'FontAwesome Icons';
+$_lang['superboxfontawesome.fontawesomePrefix'] = 'Fontawesome Prefix';
+$_lang['superboxfontawesome.fontawesomePrefix_desc'] = 'Prefix for a FontAwesome icon css class.';
+$_lang['superboxfontawesome.fontawesomeUrl'] = 'Fontawesome URL';
+$_lang['superboxfontawesome.fontawesomeUrl_desc'] = 'URL for the external FontAwesome icons sass file.';