+++ core/components/superboxfontawesome/docs/changelog.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-01-07
+
+### Added
+
+- Compatibility with SuperBoxSelect 3.0.0
+- Addidtional superseding configuration in the template variable
+
 ## [1.0.2] - 2019-04-26
 
-### Changed
+### Fixed
 
 - Fix an issue with multiple selections